added locations and items

+++ TextRPG_Assignment/Item.h
@@ -3,10 +3,11 @@
 #include<iostream>
 #include <string>
 #include <vector>
+#include <iomanip>
 
 class Item {
 public:
-	Item(std::string name, std::string description, std::string rarity, int value);
+	Item(std::string name, std::string description, std::string rarity, double value);
 protected:
 	std::string m_Name;
 	std::string m_Description;
@@ -16,7 +17,7 @@ protected:
 
 class Weapon: public Item {
 public:
-	Weapon(std::string name, std::string description, std::string rarity, int value, int toHitBonus, int damage);
+	Weapon(std::string name, std::string description, std::string rarity, double value, int toHitBonus, int damage);
 private:
 	int m_ToHitBonus;
 	int m_AttackDamage;
@@ -24,7 +25,7 @@ private:
 
 class Armor : public Item {
 public:
-	Armor(std::string name, std::string description, std::string rarity, int value, int acBonus, int strRequirement);
+	Armor(std::string name, std::string description, std::string rarity, double value, int acBonus, int strRequirement);
 private:
 	int m_AcBonus;
 	int m_StrengthRequirement;
@@ -32,14 +33,14 @@ private:
 
 class HealingPotion : public Item {
 public:
-	HealingPotion(std::string name, std::string description, std::string rarity, int value, int heal);
+	HealingPotion(std::string name, std::string description, std::string rarity, double value, int heal);
 private:
 	int m_HpHealAmount;
 };
 
 class UseableItems : public Item {
 public:
-	UseableItems(std::string name, std::string description, std::string rarity, int value, std::string use);
+	UseableItems(std::string name, std::string description, std::string rarity, double value, std::string use);
 private:
 	std::string m_Use;
 };
